refactor(store-core): extract conversation handlers into handlers/conversation.ts (audit P2-3)
Move the conversation family (handleConversationId, handleConversationsList,
handleHistoryReplayStart, handleHistoryReplayEnd + their payload interfaces) out
of the handlers/index.ts god-module into handlers/conversation.ts, behind the
barrel's export * re-export so the public surface is unchanged.

Drop the now-unused ConversationSummary type import from index.ts (it was only
used by handleConversationsList).

Pure move — body byte-identical to origin/main. index.ts 2360 -> 2211 lines.
